+++ src/Filament/Resources/ConversationResource.php
@@ -2,11 +2,11 @@
 
 namespace Wotz\FilamentChatbot\Filament\Resources;
 
+use Filament\Infolists\Components\Section;
 use Filament\Infolists\Components\TextEntry;
 use Filament\Infolists\Components\ViewEntry;
+use Filament\Infolists\Infolist;
 use Filament\Resources\Resource;
-use Filament\Schemas\Components\Section;
-use Filament\Schemas\Schema;
 use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Table;
 use Illuminate\Database\Eloquent\Builder;
@@ -18,7 +18,7 @@ class ConversationResource extends Resource
 {
     protected static ?string $model = AgentConversation::class;
 
-    protected static string|\BackedEnum|null $navigationIcon = 'heroicon-o-chat-bubble-left-right';
+    protected static ?string $navigationIcon = 'heroicon-o-chat-bubble-left-right';
 
     protected static ?string $slug = 'conversations';
 
@@ -54,10 +54,10 @@ class ConversationResource extends Resource
             ->defaultSort('updated_at', 'desc');
     }
 
-    public static function infolist(Schema $schema): Schema
+    public static function infolist(Infolist $infolist): Infolist
     {
-        return $schema
-            ->components([
+        return $infolist
+            ->schema([
                 Section::make('Conversation')
                     ->schema([
                         TextEntry::make('title'),